Remove production comments from parsing code.
They're a little annoying to maintain as we keep adding language
features. And there are enough now that having them so scattered makes
the descent logic harder to trace. Once all language features are in
place it could be useful to stick a full grammar somewhere near the
implementation to to document the internal parsing logic.